WIP addressing comments
